backend-python/requirements.txt
cryptography
Flask
Flask-CORS
psycopg2-binary
PyJWT